pyproject.toml
[project]
name = "meomory"
version = "0.1.0"
description = "Memory prosthetic for LLM state restoration"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pyarrow>=15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
